Fix an annoying Makefile error message while building
When running make (for anything from the top dir), it always complains
like that a few times (try to google this string!):

	Makefile:1279: *** mixed implicit and normal rules.

Apparently, GNU make 3.83 doesn't like a particular piece of Makefile.

Note that somehow busybox is still buildable w/o this patch, it's
probably because either its top Makefile is not being used for build.
But it is read when we run make from top dir and it spits the ugly error.

The fix is already there in the busybox git:
	https://bugs.busybox.net/show_bug.cgi?id=2323
	http://kan.gd/43n

Below is the original changelog message:

	build system: make 3.82 fix. Closes bug 2323
	Signed-off-by: Denys Vlasenko <vda.linux@googlemail.com>

Change-Id: If1cc080748bd7634812920478a70ea316149526c
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>