fix modprobe for android
modprobe typically looks for modules under a folder specific to
the kernel version that they were built against (namely
DEFAULT_MODULE_DIR/$(uname -r)), but this is broken on some devices
that simply place the modules under DEFAULT_MODULE_DIR.  This patch
tells modprobe to verify that DEFAULT_MODULE_DIR/$(uname -r) exists
prior to attempting to xchdir to it.  If the folder exists, it moves
into it and loads modules from there, if not, it remains in
DEFAULT_MODULE_DIR and attempts to load modules from there.

Change-Id: I42e8b09916928c1530a76615768913227ed76ea5